feat(mps-sync-plugin-lib): extract TreeChangeVisitor to a new class
