+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/SyncServiceImpl.kt
@@ -2,35 +2,23 @@ package org.modelix.mps.sync
 
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.diagnostic.logger
-import jetbrains.mps.extapi.model.SModelBase
-import jetbrains.mps.extapi.module.SModuleBase
 import jetbrains.mps.project.MPSProject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.runBlocking
-import org.jetbrains.mps.openapi.language.SProperty
-import org.jetbrains.mps.openapi.language.SReferenceLink
 import org.modelix.kotlin.utils.UnstableModelixFeature
-import org.modelix.model.api.IBranch
 import org.modelix.model.api.IBranchListener
 import org.modelix.model.api.INode
-import org.modelix.model.api.IReferenceLink
 import org.modelix.model.api.ITree
-import org.modelix.model.api.ITreeChangeVisitorEx
-import org.modelix.model.api.PropertyFromName
-import org.modelix.model.api.getNode
 import org.modelix.model.client2.ModelClientV2
 import org.modelix.model.client2.ReplicatedModel
 import org.modelix.model.client2.getReplicatedModel
 import org.modelix.model.lazy.BranchReference
 import org.modelix.model.mpsadapters.MPSLanguageRepository
-import org.modelix.mps.sync.mps.util.runReadBlocking
-import org.modelix.mps.sync.mps.util.runWriteActionInEDTBlocking
-import org.modelix.mps.sync.transformation.ITreeToSTreeTransformer
 import org.modelix.mps.sync.transformation.MpsToModelixMap
-import org.modelix.mps.sync.util.nodeIdAsLong
-import org.modelix.mps.sync.util.runIfAlone
+import org.modelix.mps.sync.transformation.modelixToMps.incremental.TreeChangeVisitor
+import org.modelix.mps.sync.transformation.modelixToMps.initial.ITreeToSTreeTransformer
 import java.net.ConnectException
 import java.net.URL
 import java.util.concurrent.atomic.AtomicReference
@@ -62,14 +50,13 @@ class SyncServiceImpl : SyncService {
             try {
                 log.info("Connecting to $serverURL")
                 modelClientV2.init()
-//                modelClientV2.initRepository(RepositoryId("lolwat"+(0..10).random().toString()))
             } catch (e: ConnectException) {
                 log.warn("Unable to connect: ${e.message} / ${e.cause}")
                 throw e
             }
         }
         log.info("Connection to $serverURL successful")
-        clientBindingMap[modelClientV2] = mutableListOf<BindingImpl>()
+        clientBindingMap[modelClientV2] = mutableListOf()
         return modelClientV2
     }
 
@@ -133,127 +120,31 @@ class SyncServiceImpl : SyncService {
 
 @UnstableModelixFeature(reason = "The new mod elix MPS plugin is under construction", intendedFinalization = "2024.1")
 class BindingImpl(
-    val replicatedModel: ReplicatedModel,
+    private val replicatedModel: ReplicatedModel,
     mpsProject: MPSProject,
-    private val isSynchronizing: AtomicReference<Boolean>,
-    private val nodeMap: MpsToModelixMap,
+    isSynchronizing: AtomicReference<Boolean>,
+    nodeMap: MpsToModelixMap,
 ) : IBinding {
 
-    private val branch: IBranch = replicatedModel.getBranch()
-    private val modelAccess = mpsProject.modelAccess
-
     init {
-        setupBinding()
-    }
-
-    private fun setupBinding() {
-        // listen to changes on the branch in the replicatedModel (model-server side) ...
-        // TODO refactor the IBranchListener to a new class
-        branch.addListener(object : IBranchListener {
-            fun handleThrowable(t: Throwable) {
-                t.printStackTrace()
-            }
-
+        replicatedModel.getBranch().addListener(object : IBranchListener {
             override fun treeChanged(oldTree: ITree?, newTree: ITree) {
-                if (oldTree == null) {
-                    return
+                if (oldTree != null) {
+                    newTree.visitChanges(
+                        oldTree,
+                        TreeChangeVisitor(replicatedModel, mpsProject, isSynchronizing, nodeMap),
+                    )
                 }
-                newTree.visitChanges(
-                    oldTree,
-                    object : ITreeChangeVisitorEx {
-
-                        fun getBranch() = replicatedModel.getBranch()
-
-                        override fun referenceChanged(nodeId: Long, role: String) {
-                            // TODO how to transform modelImport, modelDependency, languageDependency changes?
-                            // TODO Are they also "referenceChanged" events?
-                            isSynchronizing.runIfAlone(::handleThrowable) {
-                                val sNode = nodeMap.getNode(nodeId)!!
-                                var sReferenceLink: SReferenceLink? = null
-                                modelAccess.runReadBlocking {
-                                    sReferenceLink = sNode.concept.referenceLinks.find { it.name == role }
-                                }
-
-                                val iNode = getBranch().getNode(nodeId)
-                                var iReferenceLink: IReferenceLink?
-                                var targetINode: INode? = null
-                                getBranch().runRead {
-                                    iReferenceLink = iNode.getReferenceLinks().find { it.getSimpleName() == role }
-                                    targetINode = iReferenceLink?.let { iNode.getReferenceTarget(it) }
-                                }
-                                val targetSNode = targetINode?.let { nodeMap.getNode(it.nodeIdAsLong()) }
-
-                                modelAccess.runWriteActionInEDTBlocking {
-                                    sNode.setReferenceTarget(sReferenceLink!!, targetSNode)
-                                }
-                            }
-                        }
-
-                        override fun propertyChanged(nodeId: Long, role: String) {
-                            // TODO what if nodeId is a Model or a Module?
-                            isSynchronizing.runIfAlone(::handleThrowable) {
-                                val sNode = nodeMap.getNode(nodeId)!!
-                                var sProperty: SProperty? = null
-                                modelAccess.runReadBlocking {
-                                    sProperty = sNode.concept.properties.find { it.name == role }
-                                }
-
-                                val iNode = getBranch().getNode(nodeId)
-                                val iProperty = PropertyFromName(role)
-                                var value: String? = null
-                                getBranch().runRead { value = iNode.getPropertyValue(iProperty) }
-
-                                modelAccess.runWriteActionInEDTBlocking {
-                                    sNode.setProperty(sProperty!!, value)
-                                }
-                            }
-                        }
-
-                        override fun nodeRemoved(nodeId: Long) {
-                            val sNode = nodeMap.getNode(nodeId)
-                            sNode?.let {
-                                modelAccess.runWriteActionInEDTBlocking {
-                                    it.delete()
-                                    nodeMap.remove(nodeId)
-                                }
-                            }
-
-                            val sModel = nodeMap.getModel(nodeId)
-                            sModel?.let {
-                                val sModule = sModel.module
-                                require(sModel is SModelBase) { "Model ${sModel.modelId} is not SModelBase" }
-                                require(sModule is SModuleBase) { "Module ${sModule.moduleId} is not SModuleBase" }
-                                sModule.unregisterModel(sModel)
-                                nodeMap.remove(nodeId)
-                            }
-
-                            val sModule = nodeMap.getModule(nodeId)
-                            sModule?.let {
-                                require(sModule is SModuleBase) { "Module ${sModule.moduleId} is not SModuleBase" }
-                                sModule.dispose()
-                                nodeMap.remove(nodeId)
-                            }
-                        }
-
-                        override fun nodeAdded(nodeId: Long) {
-                            // TODO decide if node is model, module or model and transform using the corresponding transformation methods
-                            println("node added $nodeId")
-                        }
-
-                        override fun childrenChanged(nodeId: Long, role: String?) {}
-
-                        override fun containmentChanged(nodeId: Long) {}
-                    },
-                )
             }
         })
     }
 
-    override fun deactivate(callback: Runnable?) {
-        replicatedModel.dispose()
-    }
-
     override fun activate(callback: Runnable?) {
         TODO("Not yet implemented")
+    }
+
+    override fun deactivate(callback: Runnable?) {
+        // TODO unregister Model/Module/NodeChangeListeners?
+        replicatedModel.dispose()
     }
 }
