+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/transformation/mpsToModelix/incremental/NodeChangeListener.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.mps.listeners
+package org.modelix.mps.sync.transformation.mpsToModelix.incremental
 
 import org.jetbrains.mps.openapi.event.SNodeAddEvent
 import org.jetbrains.mps.openapi.event.SNodeRemoveEvent
